+++ Chartboost/ChartboostAdapterConfiguration.h
@@ -30,7 +30,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)initializeNetworkWithConfiguration:(NSDictionary<NSString *, id> * _Nullable)configuration
                                   complete:(void(^ _Nullable)(NSError * _Nullable))complete;
 
-+ (NSString*)mediationString;
++ (NSString *)adapterVersion;
 
 @end